requirements.txt
Flask==1.0.2
Flask-SocketIO==3.1.2
paho-mqtt==1.5.0
minio==5.0.5